+++ cmd/gogoodwe/version.txt
@@ -1,1 +1,1 @@
-v2.0.5
+v2.0.6
